Extract hero and app sections into components
Move large hero and app markup out of the landing-page shell into dedicated components for better encapsulation and readability. Added ab-hero-section (HTML/SCSS/TS) and ab-app-section (HTML/SCSS/TS) with OnPush change detection; hero now owns the stats data and imports RouterLink. Updated the landing-page template to use <ab-hero-section /> and <ab-app-section /> and updated its component imports (removed the inline stats array). Also moved a mobile stats spacing rule out of the global _landing.scss into the hero component SCSS and removed the duplicate global rule. Overall: componentized hero/app, cleaned up landing-page, and scoped related styles.
