+++ libs/adrian-badilla/frontend/landing-page/src/lib/components/app-section/app-section.component.html
@@ -1,0 +1,45 @@
+<section class="ab-app ab-label" data-scrub data-ss="0.92,1" data-so="0,1" style="max-width:1500px; margin:0 auto clamp(56px,8vh,120px); padding:clamp(68px,9vw,150px) clamp(32px,6vw,120px); border-radius:26px; border:1px solid rgba(255,255,255,.1); position:relative; overflow:hidden; box-shadow:0 60px 130px -50px rgba(0,0,0,.85);">
+  <div style="position:absolute; top:0; left:0; right:0; height:1px; background:linear-gradient(90deg, transparent, hsl(356 88% 53% / .55), transparent); z-index:1;"></div>
+  <div class="ab-label__sheen" aria-hidden="true"></div>
+  <div class="ab-app__grid" style="position:relative; z-index:1; display:grid; grid-template-columns:1fr 1fr; gap:clamp(40px,5.5vw,96px); align-items:center;">
+    <div>
+      <div style="display:flex; align-items:center; gap:16px; margin-bottom:14px;">
+        <span class="ab-chrome" style="font-family:'Anton',sans-serif; font-size:1.4rem;">05</span>
+        <span style="display:inline-flex; align-items:center; gap:9px;">
+          <span style="width:8px; height:8px; border-radius:50%; background:#e20613; box-shadow:0 0 12px rgba(226,6,19,.85); animation:ab-pulse 2s ease-in-out infinite;"></span>
+          <span style="font-size:13px; font-weight:700; letter-spacing:.22em; text-transform:uppercase; color:var(--accent, hsl(356 88% 53%));">Próximamente</span>
+        </span>
+      </div>
+      <div class="ab-label__eyebrow ab-chrome" style="margin-bottom:14px;">Entreno · Dieta · Seguimiento</div>
+      <h2 style="font-family:'Anton',sans-serif; font-weight:400; font-size:clamp(3rem,6.4vw,5.4rem); line-height:.92; text-transform:uppercase; letter-spacing:-.01em; margin:0 0 24px;"><span class="ab-foil">Toda tu rutina,</span><br><span style="color:transparent; -webkit-text-stroke:clamp(1.2px,.16vw,2.4px) var(--accent, hsl(356 88% 53%));">en una app</span></h2>
+      <p style="font-size:1.28rem; line-height:1.6; color:rgba(255,255,255,.74); max-width:580px; margin:0 0 40px;">Calendario de entrenos, rutinas personalizadas, control de calorías y notificaciones para mantenerte al día. Disponible muy pronto.</p>
+      <div style="display:flex; gap:14px; flex-wrap:wrap;">
+        <div class="ab-store" aria-label="Próximamente en App Store">
+          <svg viewBox="0 0 24 24" fill="currentColor" aria-hidden="true"><path d="M17.05 12.74c-.02-2.18 1.78-3.23 1.86-3.28-1.01-1.48-2.59-1.69-3.15-1.71-1.34-.13-2.62.79-3.3.79-.68 0-1.73-.77-2.84-.75-1.46.02-2.81.85-3.56 2.16-1.52 2.64-.39 6.55 1.09 8.69.72 1.05 1.58 2.22 2.71 2.18 1.09-.04 1.5-.7 2.82-.7 1.31 0 1.68.7 2.83.68 1.17-.02 1.91-1.06 2.62-2.12.83-1.22 1.17-2.4 1.19-2.46-.03-.01-2.28-.87-2.31-3.47zM14.93 5.86c.6-.74 1.01-1.76.9-2.78-.87.04-1.93.58-2.56 1.31-.56.65-1.05 1.7-.92 2.7.97.08 1.97-.5 2.58-1.23z"/></svg>
+          <span><small>Próximamente en</small><strong>App Store</strong></span>
+        </div>
+        <div class="ab-store" aria-label="Próximamente en Google Play">
+          <svg viewBox="0 0 24 24" fill="currentColor" aria-hidden="true"><path d="M3.6 2.3c-.25.25-.4.63-.4 1.12v17.16c0 .49.15.87.4 1.12l.06.05 9.6-9.6v-.23l-9.6-9.6-.06.04zM16.7 15.38l-3.2-3.2v-.23l3.2-3.21.07.04 3.79 2.16c1.08.61 1.08 1.62 0 2.24l-3.79 2.16-.07.04zM16.77 15.34 13.5 12.07 3.6 21.97c.36.38.95.43 1.61.05l11.56-6.68M16.77 8.66 5.21 2.04c-.66-.39-1.25-.34-1.61.04l9.9 9.9 3.27-3.32z"/></svg>
+          <span><small>Próximamente en</small><strong>Google Play</strong></span>
+        </div>
+      </div>
+    </div>
+    <div style="position:relative; display:flex; justify-content:center;">
+      <div style="position:absolute; inset:0; background:radial-gradient(circle at 50% 45%, hsl(356 88% 53% / .15), transparent 60%); filter:blur(6px); pointer-events:none;"></div>
+      <img src="/global/assets/img/celphone-Image.webp" alt="App Adrián Badilla" loading="lazy" style="position:relative; width:100%; max-width:500px; height:auto; object-fit:contain; filter:drop-shadow(0 36px 64px rgba(0,0,0,.62));">
+    </div>
+  </div>
+  <!-- supplement-label stat band: app features as "nutrition facts" -->
+  <div class="ab-label__band">
+    <span class="ab-label__benefit">Tu progreso, bajo control</span>
+    <div class="ab-label__cells">
+      <div class="ab-label__cell"><b>100%</b><small>Personalizado</small></div>
+      <div class="ab-label__cell"><b>24/7</b><small>Seguimiento</small></div>
+      <div class="ab-label__cell"><b>0</b><small>Excusas</small></div>
+    </div>
+    <div class="ab-label__tested">
+      <svg viewBox="0 0 24 24" fill="none" aria-hidden="true"><path d="M12 2.5l2.6 1.9 3.2-.1 1 3.05 2.6 1.85-1 3.05 1 3.05-2.6 1.85-1 3.05-3.2-.1L12 21.5l-2.6-1.9-3.2.1-1-3.05L2.6 14.8l1-3.05-1-3.05 2.6-1.85 1-3.05 3.2.1L12 2.5z" fill="rgba(255,255,255,.16)" stroke="currentColor" stroke-width="1.1" stroke-linejoin="round"/><path d="M8.4 12.2l2.4 2.4 4.8-5" stroke="currentColor" stroke-width="1.8" stroke-linecap="round" stroke-linejoin="round"/></svg>
+      <span>Respaldado por +30 campeones</span>
+    </div>
+  </div>
+</section>
